fix: reuse code to avoid duplication (code climate)
